move api clients and rename
