+++ frontend/components/Domain/Recipe/RecipeAssets.vue
@@ -71,7 +71,7 @@
 </template>
 
 <script>
-import { useApiSingleton } from "~/composables/use-api";
+import { useUserApi } from "~/composables/api";
 export default {
   props: {
     slug: {
@@ -88,7 +88,7 @@ export default {
     },
   },
   setup() {
-    const api = useApiSingleton();
+    const api = useUserApi();
 
     return { api };
   },
