+++ frontend/components/Domain/Recipe/RecipeCategoryTagSelector.vue
@@ -44,7 +44,7 @@
 
 <script>
 import RecipeCategoryTagDialog from "./RecipeCategoryTagDialog";
-import { useApiSingleton } from "~/composables/use-api";
+import { useUserApi } from "~/composables/api";
 import { useTags, useCategories } from "~/composables/use-tags-categories";
 const MOUNTED_EVENT = "mounted";
 export default {
@@ -91,7 +91,7 @@ export default {
   },
 
   setup() {
-    const api = useApiSingleton();
+    const api = useUserApi();
 
     const { allTags, useAsyncGetAll: getAllTags } = useTags();
     const { allCategories, useAsyncGetAll: getAllCategories } = useCategories();
