+++ frontend/components/Domain/Recipe/RecipeImageUploadBtn.vue
@@ -41,7 +41,7 @@
 
 <script>
 import { defineComponent } from "@nuxtjs/composition-api";
-import { useApiSingleton } from "~/composables/use-api";
+import { useUserApi } from "~/composables/api";
 const REFRESH_EVENT = "refresh";
 const UPLOAD_EVENT = "upload";
 export default defineComponent({
@@ -52,7 +52,7 @@ export default defineComponent({
     },
   },
   setup() {
-    const api = useApiSingleton();
+    const api = useUserApi();
 
     return { api };
   },
