+++ frontend/composables/api/index.ts
@@ -1,1 +1,2 @@
 export { useStaticRoutes } from "./static-routes";
+export { useAdminApi, useUserApi } from "./api-client";
