+++ frontend/composables/use-uuid.ts
@@ -1,9 +1,3 @@
-const max = 1000000;
-
-export function uniqueId() {
-  return Date.now() + Math.random() * max;
-}
-
 export function uuid4() {
   // @ts-ignore
   return ([1e7] + -1e3 + -4e3 + -8e3 + -1e11).replace(/[018]/g, (c) =>
