+++ frontend/pages/admin/parser.vue
@@ -69,12 +69,12 @@
 <script lang="ts">
 import { defineComponent, reactive, ref, toRefs } from "@nuxtjs/composition-api";
 import { Confidence, Parser } from "~/api/class-interfaces/recipes";
-import { useApiSingleton } from "~/composables/use-api";
+import { useUserApi } from "~/composables/api";
 
 export default defineComponent({
   layout: "admin",
   setup() {
-    const api = useApiSingleton();
+    const api = useUserApi();
 
     const state = reactive({
       loading: false,
